backend/requirements.txt
google-generativeai==0.4.1
langchain==0.2.11
langchain-core==0.2.23
langchain-community==0.2.10
sentence-transformers==2.6.1
transformers==4.36.2
huggingface-hub==0.20.3
scikit-learn==1.5.0
pypdf==4.3.0
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.24.0
streamlit==1.32.0
packaging==23.2
pydantic==2.7.4
pydantic-settings==2.7.1
urllib3==2.3.0
